+++ src/diffusers/models/embeddings.py
@@ -263,12 +263,8 @@ class PositionalEmbedding(nn.Module):
 
     def __init__(self, embed_dim: int, max_seq_length: int = 32):
         super().__init__()
-        position = torch.arange(max_seq_length).unsqueeze(1)
-        div_term = torch.exp(torch.arange(0, embed_dim, 2) * (-math.log(10000.0) / embed_dim))
-
-        pe = torch.zeros(1, max_seq_length, embed_dim)
-        pe[0, :, 0::2] = torch.sin(position * div_term)
-        pe[0, :, 1::2] = torch.cos(position * div_term)
+        pe = get_1d_sincos_pos_embed_from_grid(embed_dim, torch.arange(32))
+        pe = torch.tensor(pe, dtype=torch.float32).unsqueeze(0)
         self.register_buffer("pe", pe)
 
     def forward(self, x):
